% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BANN.R
\name{BANN}
\alias{BANN}
\title{BANN main function}
\usage{
BANN(X, mask, y, centered = FALSE, numModels = 20, tol = 1e-04,
  maxiter = 10000, show_progress = TRUE)
}
\description{
BANN main function
}
\examples{
res = BANN(X, mask, y)
}
