% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BANNvarEM.R
\name{BANNvarEM}
\alias{BANNvarEM}
\title{BANN function for free parameters and variance parameters updates.}
\usage{
BANNvarEM(X, y, centered = FALSE, numModels = 20, tol = 1e-04,
  maxiter = 10000, show_progress = TRUE)
}
\description{
BANN function for free parameters and variance parameters updates.
}
