% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePVE.R
\name{estimatePVE}
\alias{estimatePVE}
\title{BANN function for estimating proportiona of variance explained (PVE).}
\usage{
estimatePVE(fit, X, nr = 1000)
}
\description{
BANN function for estimating proportiona of variance explained (PVE).
}
