% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innerloop.R
\name{innerloop}
\alias{innerloop}
\title{BANN function for inner loop within each one model to update free parameters and variance parameters.}
\usage{
innerloop(X, y, xy, d, tau, sigma, logodds, alpha, mu, update.order,
  tol = 1e-04, maxiter = 10000, outer.iter = NULL)
}
\description{
BANN function for inner loop within each one model to update free parameters and variance parameters.
}
