% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outerloop.R
\name{outerloop}
\alias{outerloop}
\title{BANN function for outer loop of all models to update free parameters and variance parameters.}
\usage{
outerloop(X, I, y, xy, d, SIy, SIX, tau, sigma, logodds, alpha, mu,
  update.order, tol, maxiter, outer.iter)
}
\description{
BANN function for outer loop of all models to update free parameters and variance parameters.
}
