% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varamParamUpdate.R
\name{varParamUpdate}
\alias{varParamUpdate}
\title{BANN function for updating free parameters and variance parameters.}
\usage{
varParamUpdate(X, tau, sigma, logodds, xy, d, alpha0, mu0, Xr0, updates)
}
\description{
BANN function for updating free parameters and variance parameters.
}
